+++ shared/editor/nodes/Emoji.tsx
@@ -7,39 +7,14 @@ import {
 } from "prosemirror-model";
 import { Command, TextSelection } from "prosemirror-state";
 import { Primitive } from "utility-types";
-import Suggestion from "../extensions/Suggestion";
+import Extension from "../lib/Extension";
 import { getEmojiFromName } from "../lib/emoji";
 import { MarkdownSerializerState } from "../lib/markdown/serializer";
-import { SuggestionsMenuType } from "../plugins/Suggestions";
 import emojiRule from "../rules/emoji";
 
-/**
- * Languages using the colon character with a space in front in standard
- * punctuation. In this case the trigger is only matched once there is additional
- * text after the colon.
- */
-const languagesUsingColon = ["fr"];
-
-export default class Emoji extends Suggestion {
+export default class Emoji extends Extension {
   get type() {
     return "node";
-  }
-
-  get defaultOptions() {
-    const languageIsUsingColon =
-      typeof window === "undefined"
-        ? false
-        : languagesUsingColon.includes(window.navigator.language.slice(0, 2));
-
-    return {
-      type: SuggestionsMenuType.Emoji,
-      openRegex: new RegExp(
-        `(?:^|\\s):([0-9a-zA-Z_+-]+)${languageIsUsingColon ? "" : "?"}$`
-      ),
-      closeRegex:
-        /(?:^|\s):(([0-9a-zA-Z_+-]*\s+)|(\s+[0-9a-zA-Z_+-]+)|[^0-9a-zA-Z_+-]+)$/,
-      enabledInTable: true,
-    };
   }
 
   get name() {
